requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
autopep8==2.0.1
boto3==1.26.41
botocore==1.29.41
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.0.1
cryptography==39.0.0
distlib==0.3.6
Django==4.1.3
django-cors-headers==3.11.0
django-model-utils==4.3.1
django-ses==3.5.0
djangorestframework==3.14.0
filelock==3.9.0
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
identify==2.5.19
idna==3.4
jmespath==1.0.1
jsonwebtoken==1.0.0
Naked==0.1.32
nodeenv==1.7.0
numpy==1.24.1
Pillow==9.4.0
platformdirs==3.1.0
pre-commit==2.19.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
PyYAML @ file:///private/tmp/pyyaml-20230717-6270-1iiwi1j/PyYAML-6.0.1
redis==4.5.4
requests==2.28.2
s3transfer==0.6.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
urllib3==1.26.13
virtualenv==20.21.1
zope.event==4.6
zope.interface==6.0
